fix: Move KtSchemaReflectionFactory to separate package
