+++ tachyon-kotlin-kt-schema/pom.xml
@@ -17,7 +17,7 @@
 
     <name>Tachyon MCP Kotlin kt-schema</name>
     <description>
-        kt-schema integration for Tachyon MCP Kotlin module with reflection JsonSchema generator.
+        Experimental kt-schema integration for Tachyon MCP Kotlin module with reflection JsonSchema generator.
     </description>
 
     <dependencies>
