+++ tachyon-kotlin-kt-schema/src/main/kotlin/dev/tachyonmcp/kotlin/server/json/ktschema/KtSchemaReflectionFactory.kt
@@ -1,5 +1,7 @@
-// Copyright (c) 2026 Konstantin Pavlov/IT Staff and contributors.
-package dev.tachyonmcp.kotlin.server.json
+/*
+ * Copyright (c) 2026 Konstantin Pavlov/IT Staff and contributors.
+ */
+package dev.tachyonmcp.kotlin.server.json.ktschema
 
 import dev.tachyonmcp.api.annotations.ExperimentalApi
 import dev.tachyonmcp.api.json.JsonSchema
