+++ tachyon-kotlin-kt-schema/src/main/resources/META-INF/services/dev.tachyonmcp.api.json.spi.JsonSchemaFactory
@@ -1,1 +1,1 @@
-dev.tachyonmcp.kotlin.server.json.KtSchemaReflectionFactory
+dev.tachyonmcp.kotlin.server.json.ktschema.KtSchemaReflectionFactory
