+++ tachyon-kotlin-kt-schema/src/test/kotlin/dev/tachyonmcp/kotlin/server/json/ktschema/KtSchemaReflectionFactoryTest.kt
@@ -1,8 +1,11 @@
-// Copyright (c) 2026 Konstantin Pavlov/IT Staff and contributors.
-package dev.tachyonmcp.kotlin.server.json
+/*
+ * Copyright (c) 2026 Konstantin Pavlov/IT Staff and contributors.
+ */
+package dev.tachyonmcp.kotlin.server.json.ktschema
 
 import dev.tachyonmcp.api.json.JsonSchema
 import dev.tachyonmcp.api.json.spi.JsonSchemaFactory
+import dev.tachyonmcp.core.server.json.KtSchemaResourceFactory
 import io.kotest.assertions.json.shouldEqualJson
 import io.kotest.matchers.shouldBe
 import kotlinx.serialization.SerialName
@@ -11,7 +14,7 @@ import org.junit.jupiter.api.Test
 import java.util.ServiceLoader
 
 /**
- * Verifies [KtSchemaReflectionFactory] is discovered through [ServiceLoader] and that
+ * Verifies [dev.tachyonmcp.kotlin.server.json.ktschema.KtSchemaReflectionFactory] is discovered through [ServiceLoader] and that
  * `JsonSchema.generated(Class)` — the reified `typedTool` path — resolves a kt-schema-generated
  * schema via the reflection factory whenever no build-time codegen resource exists. This is the
  * dedicated `tachyon-kotlin-kt-schema` integration artifact, so the provider must always be
@@ -34,8 +37,8 @@ internal class KtSchemaReflectionFactoryTest {
 
         factories.map { it.javaClass.name }.filter { it.contains("KtSchema") } shouldBe
             listOf(
-                "dev.tachyonmcp.core.server.json.KtSchemaResourceFactory",
-                "dev.tachyonmcp.kotlin.server.json.KtSchemaReflectionFactory",
+                KtSchemaResourceFactory::class.java.canonicalName,
+                KtSchemaReflectionFactory::class.java.canonicalName,
             )
     }
